scrapper/requirements.txt
requests >= 2.25.1
beautifulsoup4 >= 4.9.3
python-dateutil >= 2.8.1

flask >= 2.0.1
python-dotenv >= 0.17.1
